perf: cache alias mapping results
I can't think of a reason not to cache these. I would have preferred to have done all mapping in memory (perhaps for data sets under a certain size) but that would be a much more invasive change.

This relatively minor fix brings down an import job that takes us ~25 minutes to a more reasonable ~5 minutes. It's mapping ~6000 records against a relatively small dataset (systemusers). All reference data is assigned to the same mapped user, so it makes 6000 requests to get the same user. I imagine this is not an uncommon scenario.