perf: cache alias mapping results (#59)
Caches GUIDS for alias mappings to cut down on repeat API requests.

This relatively minor fix brings down an import job that takes us 25 minutes to a more reasonable 5 minutes. It's mapping 6000 records against a relatively small dataset (systemusers). All reference data is assigned to the same mapped user, so it makes 6000 requests to get the same user. This commit ensures that only one request is ever made for that same set of aliased values.